[fix] : 에러 수정

+++ backend/src/main/java/com/hans/hans/domain/mode/repository/ModeRepository.java
@@ -4,6 +4,6 @@ import com.hans.hans.domain.mode.entity.Mode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ModeRepository extends JpaRepository<Mode,Long> {
-    Mode findByModeSequence(Long modeSequence);
+    Mode findByModeSequence(int modeSequence);
     Mode findByModeName(String modeName);
 }
