+++ backend/src/main/java/com/hans/hans/domain/ranking/service/RankingService.java
@@ -6,7 +6,7 @@ import com.hans.hans.domain.ranking.dto.RankingResponseDto;
 import org.springframework.data.domain.Pageable;
 
 public interface RankingService {
-    RankingResponseDto getRankingListByMode(Long id, Pageable pageable);
+    RankingResponseDto getRankingListByMode(int id, Pageable pageable);
     RankingResponseByMemberDto createRanking(RankingRequestDto rankingRequestDto);
 
 }
